fix: consistent type for all implementations
The instances of every created instance should have the same type as when used through the API. Before this commit, curve25519 and secp256k1 both typed their instances as “ec” in order for the “ec” implementations to work. After this commit, the instances are both created and restored with the correct type.